requirements.txt
Django==2.1
Pillow
psycopg2-binary
django-crispy-forms==1.7.2
python-dateutil